requirements.txt
flask
requests
python-dotenv
python-bitget
pybitget
